requirements.txt
semver
GitPython